style(client): rename table folder
